requirements.txt
Django==3.0.4
django-rest-framework==0.1.0
djangorestframework==3.11.0
requests==2.22.0
scrapd==3.1.1
Scrapy==1.8.0
scrapyd-client==1.1.0
# mysqlclient==1.4.6
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
pyodbc==4.0.30
random-user-agent==1.0.1
js2xml==0.3.1